eso-prototype-server/eso-service/eso-service-rest/requirements.txt
Flask==0.12.0
flask-cors==3.0.3
jsonschema==2.6.0
requests==2.13.0
PyYAML==3.11
retrying==1.3.3
nose==1.3.7
nose2
pep8==1.7.0
pylint==1.7.2
mock==2.0.0
requests_mock==1.3.0
flask-sse==0.2.1